Game Completed (sorta)
As I was going through the game I decided that I learned what I needed to learn, what this is mostly missing is the fact that it's doesn't have music, the game end screen, and some minor interaction. It not anything major but I am feeling bored for the project. I might change my mind later but it is what it is.